Make large operators convert to the successors in display style
Finally, this is finding the first usage of the extra data in the tfm
files. When "large operators" are set in display style, they are
converted into their successor character, which displays a larger
version of that character. Using the font information, it's fairly
simple to find the successor character for a given character.

+++ src/tfm/mod.rs
@@ -12,7 +12,7 @@ struct TFMHeader {
 enum CharKind {
     Vanilla,
     LigKern { ligkern_index: usize },
-    CharList { next_char: usize },
+    CharList { next_char: u8 },
     Extensible { ext_recipe_index: usize },
 }
